moved models to higher directory, added manage.py in case migrations will be needed, redone database (again) and fixed deletions of boards
